fix and move tests
